C++ link takes years😅😅😅😅😅😅😅😅😅😅😅😅😅

+++ build_all.sh
@@ -2,14 +2,24 @@
 eval "$(conda shell.bash hook)" 
 conda activate ~/env/nccl_mod
 
+
+
 set -ex
 export CUDA_HOME=$CONDA_PREFIX
 
-cd nccl
+# return error if NCCL_BUILD_PATH is not set
+if [ -z "$NCCL_BUILD_PATH" ]; then
+    echo "NCCL_BUILD_PATH is not set"
+    exit 1
+fi
+
+rm -rf $NCCL_BUILD_PATH/nccl
+cp -r nccl $NCCL_BUILD_PATH
+cd $NCCL_BUILD_PATH/nccl
 rm -rf build
 make -j src.build
 make install PREFIX=$CONDA_PREFIX
-cd ..
+cd -
 
 # # cd nccl-tests
 # rm -rf build
